Move interfaces to separate directory
