+++ packages/framework/view-adapters/src/index.ts
@@ -3,4 +3,5 @@
  * Licensed under the MIT License.
  */
 
+export * from "./interfaces";
 export * from "./react";
