+++ packages/framework/view-adapters/src/interfaces/interfaces.ts
