+++ packages/framework/view-adapters/src/react/embeddedComponent.tsx
@@ -5,7 +5,7 @@
 
 import { IComponent, IComponentHTMLView, IComponentHTMLVisual } from "@microsoft/fluid-component-core-interfaces";
 import * as React from "react";
-import { IComponentReactViewable } from "./interfaces";
+import { IComponentReactViewable } from "../interfaces";
 
 export interface IEmbeddedComponentProps {
     component: IComponent;
